Fix policies

+++ src/Controller/ArticlesController.php
@@ -56,7 +56,12 @@ class ArticlesController extends AppController
     public function add()
     {
         $article = $this->Articles->newEmptyEntity();
-        $this->Authorization->authorize($article);
+        
+        //$this->Authorization->authorize($article);
+        if (!$this->Authorization->can($article, 'add')) {
+            $this->Flash->error(__('Restricted access.'));
+            return $this->redirect(['controller' => 'Articles', 'action' => 'index']);
+        }
         
         if ($this->request->is('post')) {
             $article = $this->Articles->patchEntity($article, $this->request->getData());
@@ -97,7 +102,11 @@ class ArticlesController extends AppController
             'contain' => ['Tags'],
         ]);
         
-        $this->Authorization->authorize($article);
+        //$this->Authorization->authorize($article);
+        if (!$this->Authorization->can($article, 'edit')) {
+            $this->Flash->error(__('Restricted access.'));
+            return $this->redirect(['controller' => 'Articles', 'action' => 'index']);
+        }
         
         if ($this->request->is(['patch', 'post', 'put'])) {
             $article = $this->Articles->patchEntity($article, $this->request->getData(), [
@@ -134,7 +143,11 @@ class ArticlesController extends AppController
         $this->request->allowMethod(['post', 'delete']);
         $article = $this->Articles->get($id);
         
-        $this->Authorization->authorize($article);
+        //$this->Authorization->authorize($article);
+        if (!$this->Authorization->can($article, 'delete')) {
+            $this->Flash->error(__('Restricted access.'));
+            return $this->redirect(['controller' => 'Articles', 'action' => 'index']);
+        }
         
         if ($this->Articles->delete($article)) {
             $this->Flash->success(__('The article has been deleted.'));
